# Improve performance of counters calculation