--- a/OrleansDashboard/DashboardCounters.cs
+++ b/OrleansDashboard/DashboardCounters.cs
@@ -1,13 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
-using Orleans.Runtime;
-using System.Collections.Concurrent;
 using System.Linq;
-using Orleans;
 
 namespace OrleansDashboard
 {
-
     public class SimpleGrainStatisticCounter
     {
         public int ActivationCount { get; set; }
@@ -32,7 +28,7 @@ namespace OrleansDashboard
         public double ElapsedTime { get; set; }
         public string GrainAndMethod => $"{this.Grain}.{this.Method}";
     }
-    
+
     public class SiloDetails
     {
         public int FaultZone { get; set; }
@@ -68,8 +64,57 @@ namespace OrleansDashboard
         public SimpleGrainStatisticCounter[] SimpleGrainStats { get; set; }
         public int TotalActivationCount { get; set; }
         public Queue<int> TotalActivationCountHistory { get; set; }
+    }
 
+    internal class GrainSiloKey
+    {
+        public GrainSiloKey(string grainName, string siloAddress)
+        {
+            GrainName = grainName;
+            SiloAddress = siloAddress;
+        }
 
+        public string GrainName { get; }
 
+        public string SiloAddress { get; }
+
+        protected bool Equals(GrainSiloKey other)
+        {
+            return string.Equals(GrainName, other.GrainName)
+                && string.Equals(SiloAddress, other.SiloAddress);
+        }
+
+        public override bool Equals(object obj)
+        {
+            if (ReferenceEquals(null, obj)) return false;
+            if (ReferenceEquals(this, obj)) return true;
+            if (obj.GetType() != this.GetType()) return false;
+            return Equals((GrainSiloKey)obj);
+        }
+
+        public override int GetHashCode()
+        {
+            unchecked
+            {
+                return (GrainName.GetHashCode() * 397) ^ SiloAddress.GetHashCode();
+            }
+        }
+
+        public static bool operator ==(GrainSiloKey left, GrainSiloKey right)
+        {
+            return Equals(left, right);
+        }
+
+        public static bool operator !=(GrainSiloKey left, GrainSiloKey right)
+        {
+            return !Equals(left, right);
+        }
+    }
+
+    internal class AggregatedGrainTotals
+    {
+        public double TotalAwaitTime { get; set; }
+        public long TotalCalls { get; set; }
+        public long TotalExceptions { get; set; }
     }
 }