--- a/PerformanceTests/Program.cs
+++ b/PerformanceTests/Program.cs
@@ -1,0 +1,12 @@
+﻿using BenchmarkDotNet.Running;
+
+namespace PerformanceTests
+{
+    internal class Program
+    {
+        private static void Main(string[] args)
+        {
+            BenchmarkRunner.Run<DashboardGrainBenchmark>();
+        }
+    }
+}