--- a/PerformanceTests/TestDashboardGrain.cs
+++ b/PerformanceTests/TestDashboardGrain.cs
@@ -1,0 +1,21 @@
+﻿using OrleansDashboard;
+using System;
+using System.Threading.Tasks;
+
+namespace PerformanceTests
+{
+    public class TestDashboardGrain : DashboardGrain
+    {
+        protected override IDisposable RegisterTimer(Func<object, Task> asyncCallback, object state, TimeSpan dueTime, TimeSpan period)
+        {
+            return new EmptyDisposable();
+        }
+
+        private class EmptyDisposable : IDisposable
+        {
+            public void Dispose()
+            {
+            }
+        }
+    }
+}