… PerformanceTests/PerformanceTests.csproj
… PerformanceTests/packages.config