We've seen certain performance degradation on the silo holding the dashboard grain to the point that we had to turn it off.

In this PR I've tried to improve performance of recalculation of grain counters in `DashboardGrain`. Very basic performance tests that added in this PR show certain improvement on my PC.

Before the code change:
```
// * Summary *

BenchmarkDotNet=v0.10.6, OS=Windows 10 Redstone 1 (10.0.14393)
Processor=Intel Core i5-4440 CPU 3.10GHz (Haswell), ProcessorCount=4
Frequency=3020347 Hz, Resolution=331.0878 ns, Timer=TSC
  [Host]     : Clr 4.0.30319.42000, 32bit LegacyJIT-v4.6.1648.0
  DefaultJob : Clr 4.0.30319.42000, 32bit LegacyJIT-v4.6.1648.0


      Method | SiloCount | GrainTypeCount | GrainMethodCount | GrainActivationPerSiloCount | GrainCallsPerActivationCount |      Mean |     Error |    StdDev |
------------ |---------- |--------------- |----------------- |---------------------------- |----------------------------- |----------:|----------:|----------:|
 Recalculate |         3 |             50 |               10 |                         100 |                         1000 |  21.44 ms | 0.1148 ms | 0.1074 ms |
 Recalculate |         3 |            100 |               10 |                         100 |                         1000 |  72.59 ms | 0.3099 ms | 0.2899 ms |
 Recalculate |         3 |            200 |               10 |                         100 |                         1000 | 262.14 ms | 0.5011 ms | 0.4688 ms |
```
After the code change:
```
// * Summary *

BenchmarkDotNet=v0.10.6, OS=Windows 10 Redstone 1 (10.0.14393)
Processor=Intel Core i5-4440 CPU 3.10GHz (Haswell), ProcessorCount=4
Frequency=3020347 Hz, Resolution=331.0878 ns, Timer=TSC
  [Host]     : Clr 4.0.30319.42000, 32bit LegacyJIT-v4.6.1648.0
  DefaultJob : Clr 4.0.30319.42000, 32bit LegacyJIT-v4.6.1648.0


      Method | SiloCount | GrainTypeCount | GrainMethodCount | GrainActivationPerSiloCount | GrainCallsPerActivationCount |       Mean |     Error |    StdDev |
------------ |---------- |--------------- |----------------- |---------------------------- |----------------------------- |-----------:|----------:|----------:|
 Recalculate |         3 |             50 |               10 |                         100 |                         1000 |   592.5 us |  3.078 us |  2.879 us |
 Recalculate |         3 |            100 |               10 |                         100 |                         1000 | 1,188.5 us |  7.666 us |  7.170 us |
 Recalculate |         3 |            200 |               10 |                         100 |                         1000 | 2,397.1 us | 17.266 us | 15.306 us |
```

Let me know if it makes sense. Also if you like, I can take out the performance tests from this PR and keep the code changes only.