@richorama 2017-05-30
This looks like an excellent improvement, thank you!